pyproject.toml
[project]
name = "everyones-first-chess-workbook"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-cloud-vision>=3.12.0",
    "pillow>=12.1.0",
    "datauri>=1.0.0",
    "mistralai>=1.10.0",
    "pymupdf>=1.26.7",
    "python-chess>=1.999",
    "python-dotenv>=1.2.1",
]
